+++ Ejercicios Java/Hoja 3/Ej 5/aCelularParalelo.java
@@ -16,7 +16,7 @@ public class aCelularParalelo extends Thread
     generations = g;
 
     for(int i = 0; i < cellN; ++i)
-      primitive[i] = Math.random()*100;
+      primitive[i] = (int)(Math.random()*100);
   }
 
   public void evolution()
